chore: check if cake version file exists before trying to delete it when running in Latest mode

+++ installation/cake.ps1
@@ -30,7 +30,9 @@ function InstallCake([string] $Version)
         $packageInfo = [xml]$wc.DownloadString("https://cakebuild.net/download/bootstrapper/packages")
         Write-Host $packageInfo.packages.package.version
 
-        Remove-Item $cakeVersionFile
+        if(Test-Path -Path $cakeVersionFile){
+            Remove-Item $cakeVersionFile
+        }
     }
     else
     {
